Fixes issue where environment variable value overwrites the value in AWS CodeDeploy Deployment Group field.
When executing a build, if an environemnt variable is specified by name in the AWS CodeDeploy Deployment Group field, at the end of the build the value of the environment variable will ovewrite the name of the environment variable in that field.

Example of the bug this commit fixes:
At the start of a build, the value of AWS CodeDeploy Deployment Group is "$DEPLOYMENT_GROUP". I then do a build where the value of the $DEPLOYMENT_GROUP environment variable is "QA". If I wait until the end of the build and come back to the Configure page, the value of the field will now be "QA".