Skip codedeploy if build was not mark as success
Before we run codedeploy if build was not mark as failure.
This meant that we deployed for ABORTED, NOT_BUILT and UNSTABLE states as well
(http://javadoc.jenkins-ci.org/hudson/model/Result.html)
Until there is a finer grained configuration option, it's safer to deploy for
SUCCESS builds only.

Fixes #61